requirements.txt
python-telegram-bot==20.7
aiohttp>=3.9.0
requests>=2.31.0
coqui-tts==0.9.0
numpy>=1.25.0
torch>=2.0.0
torchaudio>=2.0.0
pydub>=0.25.1
python-dotenv>=1.0.1